Fix memory leak while creating multiple httpClient
